Moved Climbable to api package; updated Climbable javadoc.
